refactored "load database" and "start atomdb" to an independent component for better code consistency.
