+++ das-dashboard/src/components/dashboard/MainContent/sidebar/SideBar.jsx
@@ -1,6 +1,6 @@
-import { useRef, useState } from "react";
-import { List, ListItemIcon, ListItemText, CircularProgress } from "@mui/material";
-import { Dashboard as DashboardIcon, SettingsEthernet, Polyline, PlayArrow, Stop, Storage } from "@mui/icons-material";
+import { useState } from "react";
+import { List, ListItemIcon, ListItemText } from "@mui/material";
+import { Dashboard as DashboardIcon, SettingsEthernet, Polyline } from "@mui/icons-material";
 
 import {
   SidebarContainer,
@@ -13,13 +13,11 @@ import {
   ActionDivider
 } from "./sidebar.styled";
 import { useDashboardContext } from "../../../global_providers/DashboardContextProvider";
-import { useToast } from "../../../global_providers/ToastProvider";
-import { useDialog } from "../../../global_providers/DialogProvider";
-
-import { startArchitecture, stopArchitecture, startDatabases, stopDatabases } from "../../../../api/ServicesAPI";
-import { uploadMettaFile, loadMettaFile } from "../../../../api/AtomDBAPI";
-import { extractErrorDetails } from "../../../../api/APIUtils";
 import { getInfraStatus } from "../../../../utils/serviceInventory";
+
+import { ArchitectureActionControl } from "./ArchitectureActionControl";
+import { AtomDBActionControl } from "./AtomDBActionControl";
+import { MettaLoadActionControl } from "./MettaLoadActionControl";
 
 const navigationItems = [
   { key: "servers", label: "Servers", icon: SettingsEthernet, context: "servers" },
@@ -28,7 +26,7 @@ const navigationItems = [
 
 export function SideBar() {
   const [selected, setSelected] = useState("servers");
-  const [loadingAction, setLoadingAction] = useState(null);
+  const [busyActions, setBusyActions] = useState({});
 
   const {
     setCurrentContext,
@@ -37,141 +35,14 @@ export function SideBar() {
     isSwitchingHost
   } = useDashboardContext();
 
-  const { showToast } = useToast();
-  const { showConfirm } = useDialog();
-
-  const mettaInputRef = useRef(null);
-
   const currentHost = currentMachine?.serverIp;
   const isServerOffline = !currentHost || isSwitchingHost;
   const { atomDbOnline, architectureOnline } = getInfraStatus(mergedServices);
+  const isAnyActionLoading = Object.values(busyActions).some(Boolean);
 
-  const executeAsyncAction = async (actionKey, action, successMessage, errorMessage) => {
-    try {
-      setLoadingAction(actionKey);
-      await action();
-      showToast({ message: successMessage, severity: "success" });
-    } catch (err) {
-      console.error(errorMessage, err);
-      showToast({ message: errorMessage, severity: "error", details: extractErrorDetails(err) });
-    } finally {
-      setLoadingAction(null);
-    }
+  const setActionBusy = (actionKey, busy) => {
+    setBusyActions((current) => ({ ...current, [actionKey]: busy }));
   };
-
-  const handleMettaUpload = async (event) => {
-    const file = event.target.files?.[0];
-    if (!file) return;
-
-    const hostSnapshot = currentHost;
-
-    try {
-      setLoadingAction("upload-metta");
-      const uploadResponse = await uploadMettaFile(hostSnapshot, false, file);
-
-      setLoadingAction("load-metta");
-      await loadMettaFile(hostSnapshot, uploadResponse.saved_path);
-      showToast({ message: "MeTTa database loaded successfully.", severity: "success" });
-    } catch (err) {
-      const responseData = err?.response?.data;
-
-      if (err?.response?.status === 409 && responseData?.file_path) {
-        const existingFilePath = responseData.file_path;
-
-        showConfirm({
-          title: "File already exists",
-          message: `The file "${file.name}" already exists on the server.\n\nDo you want to OVERWRITE it? (Click CANCEL to safely load the existing server file instead).`,
-          onConfirm: () => {
-            executeAsyncAction(
-              "overwrite-metta",
-              async () => {
-                const res = await uploadMettaFile(hostSnapshot, true, file);
-                await loadMettaFile(hostSnapshot, res.saved_path);
-              },
-              "MeTTa file overwritten and loaded successfully.",
-              "Failed to overwrite MeTTa file."
-            );
-          },
-          onCancel: () => {
-            executeAsyncAction(
-              "load-existing-metta",
-              async () => {
-                await loadMettaFile(hostSnapshot, existingFilePath);
-              },
-              "Existing MeTTa file loaded successfully.",
-              "Failed to load existing MeTTa file."
-            );
-          }
-        });
-      } else {
-        showToast({ message: "Failed to upload MeTTa database.", severity: "error", details: extractErrorDetails(err) });
-      }
-    } finally {
-      setLoadingAction(null);
-      if (event.target) event.target.value = "";
-    }
-  };
-
-  const handleArchitectureAction = () => {
-    if (!atomDbOnline) {
-      showToast({ message: "AtomDB must be online before starting architecture.", severity: "warning" });
-      return;
-    }
-
-    if (architectureOnline) {
-      showConfirm({
-        title: "Stop DAS Services",
-        message: "Are you sure you want to stop all DAS services?\n\nEverything will be shut down except AtomDB.",
-        onConfirm: () => executeAsyncAction(
-          "stop-architecture",
-          () => stopArchitecture(currentHost),
-          "Architecture stopped successfully.",
-          "Failed to stop architecture."
-        )
-      });
-      return;
-    }
-
-    showConfirm({
-      title: "Start Architecture",
-      message: "Do you want to start the architecture?",
-      onConfirm: () => executeAsyncAction(
-        "start-architecture",
-        () => startArchitecture(currentHost),
-        "Architecture started successfully.",
-        "Failed to start architecture."
-      )
-    });
-  };
-
-  const handleDatabaseAction = () => {
-    if (atomDbOnline) {
-      showConfirm({
-        title: "Stop AtomDB",
-        message: "Are you sure you want to stop AtomDB?\n\nAll in-memory data may be lost.",
-        onConfirm: () => executeAsyncAction(
-          "stop-database",
-          () => stopDatabases(currentHost),
-          "AtomDB stopped successfully.",
-          "Failed to stop AtomDB."
-        )
-      });
-      return;
-    }
-
-    showConfirm({
-      title: "Start AtomDB",
-      message: "Do you want to start AtomDB?",
-      onConfirm: () => executeAsyncAction(
-        "start-database",
-        () => startDatabases(currentHost),
-        "AtomDB started successfully.",
-        "Failed to start AtomDB."
-      )
-    });
-  };
-
-  const isActionLoading = ["load-metta", "overwrite-metta", "load-existing-metta", "upload-metta"].includes(loadingAction);
 
   return (
     <SidebarContainer>
@@ -213,41 +84,26 @@ export function SideBar() {
           <SectionLabel>Actions</SectionLabel>
 
           <List disablePadding>
-            <input ref={mettaInputRef} type="file" accept=".metta" hidden onChange={handleMettaUpload} />
+            <MettaLoadActionControl
+              isServerOffline={isServerOffline}
+              disabled={isAnyActionLoading && !busyActions.metta}
+              onBusyChange={(busy) => setActionBusy("metta", busy)}
+            />
 
-            <StyledItem
-              disabled={!!loadingAction || isServerOffline}
-              onClick={() => !loadingAction && !isServerOffline && mettaInputRef.current?.click()}
-            >
-              <ListItemIcon>
-                {isActionLoading ? <CircularProgress size={16} /> : <Storage />}
-              </ListItemIcon>
-              <ListItemText primary="Load MeTTa Database" />
-            </StyledItem>
+            <ArchitectureActionControl
+              atomDbOnline={atomDbOnline}
+              architectureOnline={architectureOnline}
+              isServerOffline={isServerOffline}
+              disabled={isAnyActionLoading && !busyActions.architecture}
+              onBusyChange={(busy) => setActionBusy("architecture", busy)}
+            />
 
-            <StyledItem
-              disabled={!!loadingAction || isServerOffline || (!atomDbOnline && !architectureOnline)}
-              onClick={() => !loadingAction && !isServerOffline && (atomDbOnline || architectureOnline) && handleArchitectureAction()}
-            >
-              <ListItemIcon>
-                {["start-architecture", "stop-architecture"].includes(loadingAction) ? (
-                  <CircularProgress size={16} />
-                ) : architectureOnline ? <Stop /> : <PlayArrow />}
-              </ListItemIcon>
-              <ListItemText primary={architectureOnline ? "Stop Architecture" : "Start Architecture"} />
-            </StyledItem>
-
-            <StyledItem
-              disabled={!!loadingAction || isServerOffline}
-              onClick={() => !loadingAction && !isServerOffline && handleDatabaseAction()}
-            >
-              <ListItemIcon>
-                {["start-database", "stop-database"].includes(loadingAction) ? (
-                  <CircularProgress size={16} />
-                ) : atomDbOnline ? <Stop /> : <PlayArrow />}
-              </ListItemIcon>
-              <ListItemText primary={atomDbOnline ? "Stop AtomDB" : "Start AtomDB"} />
-            </StyledItem>
+            <AtomDBActionControl
+              atomDbOnline={atomDbOnline}
+              isServerOffline={isServerOffline}
+              disabled={isAnyActionLoading && !busyActions.atomdb}
+              onBusyChange={(busy) => setActionBusy("atomdb", busy)}
+            />
           </List>
         </StyledList>
       </SidebarListContainer>
